% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quebec-hex-ridings-package.R
\docType{package}
\name{quebec-hex-ridings}
\alias{quebec-hex-ridings}
\alias{quebec-hex-ridings-package}
\title{...}
\description{
\itemize{
\item URL: \url{https://gitlab.com/hrbrmstr/quebec-hex-ridings}
\item BugReports: \url{https://gitlab.com/hrbrmstr/quebec-hex-ridings/issues}
}
}
\author{
Bob Rudis (bob@rud.is)
}
